+++ config/routes.rb
@@ -6,7 +6,8 @@ Rails.application.routes.draw do
   get 'welcome', to: 'sessions#welcome'
   get 'authorized', to: 'sessions#page_requires_login'
   put '/posts/:id/like', to: 'posts#like', as: 'like'
-  
+  get 'posts/:user_id/posts', to: 'posts#showSpecificUsersPost'
+
   resources :users, only: [:new, :create]
   resources :posts
 end